LayoutTests/media/modern-media-controls/forward-button/forward-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>ForwardButton</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const forwardButton = new ForwardButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("forwardButton.element.localName", "button");
shouldBeTrue('forwardButton.element.classList.contains("icon")');
shouldBeTrue('forwardButton.element.classList.contains("forward")');
shouldBeEqualToString("forwardButton.iconName", `${Icons.Forward}`);

scheduler.frameDidFire = function()
{
    if (forwardButton.width == 0)
        return;

    shouldBeTrue('forwardButton.element.style.webkitMaskImage.includes("macOS/forward@")');
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
